fix: consolidate APIKey into api/models as single canonical definition
